refactor: Implement interface and abstract classes
Break apart shared functionality and contract requirements for future StringManipulation classes.
