+++ Shared/Helpers/StringManipulator/BaseStringManipulator.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: f64bf9d29f0b4ce4d8eb33987a3790d7
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
