+++ Shared/Helpers/StringManipulator/IStringManipulator.cs
@@ -5,9 +5,8 @@
   Copyright:      ©2023 AlchemicalFlux. All rights reserved.
 
   Last commit by: alchemicalflux 
-  Last commit at: 2023-10-17 13:42:06 
+  Last commit at: 2023-10-20 07:36:29 
 ------------------------------------------------------------------------------*/
-using System;
 using System.Collections.Generic;
 
 namespace AlchemicalFlux.Utilities.Helpers
@@ -15,37 +14,15 @@ namespace AlchemicalFlux.Utilities.Helpers
     /// <summary>
     /// Interface for string manipulation and any shared functionality.
     /// </summary>
-    public abstract class IStringManipulator
+    public interface IStringManipulator
     {
         /// <summary>
         /// Replaces all keys with their respective values within the supplied text.
         /// Will attempt to match using the longest keys first.
         /// </summary>
-        /// <param name="text">Entry to be parsed for replacement.</param>
+        /// <param name="originalText">Entry to be parsed for replacement.</param>
         /// <param name="replacements">Terms to be searched(keys) and replaced(values).</param>
         /// <returns>Parsed output containing original text with replaced values.</returns>
-        public abstract string MultipleReplace(string text, Dictionary<string, string> replacements);
-
-        /// <summary>
-        /// Checks if parameters have valid values that can be processed.
-        /// </summary>
-        /// <param name="text">Entry to be parsed for replacement.</param>
-        /// <param name="replacements">Terms to be searched(keys) and replaced(values).</param>
-        protected bool AreParametersInvalid(string text, Dictionary<string, string> replacements)
-        {
-            // Handle parameter checks and short circuts.
-            if (text == null || replacements == null)
-            {
-                var paramName = (text == null) ? nameof(text) : string.Empty;
-                paramName += (paramName != string.Empty && replacements == null) ? " and " : string.Empty;
-                paramName += (replacements == null) ? nameof(replacements) : string.Empty;
-
-                var message = "The '" + paramName + "' parameter(s) cannot be null.";
-
-                throw new ArgumentNullException(paramName, message);
-            }
-
-            return text == string.Empty || replacements.Count == 0;
-        }
+        string MultipleReplace(string originalText, Dictionary<string, string> replacements);
     }
 }
